requirements.txt
argon2-cffi==19.1.0
awsebcli==3.19.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
boto3==1.15.6
botocore==1.17.63
bs4==0.0.1
cement==2.8.2
certifi==2020.6.20
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
Django==2.1.7
django-extensions==2.1.6
django-taggit==1.3.0
docutils==0.15.2
future==0.16.0
graphviz==0.10.1
idna==2.7
jmespath==0.10.0
pathspec==0.5.9
Pillow==5.4.1
psycopg2==2.8.6
pycparser==2.19
pydot==1.4.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.3.1
requests==2.20.1
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
soupsieve==1.8
termcolor==1.1.0
urllib3==1.24.3
wcwidth==0.1.9
werkzeug==0.15.2
wincertstore==0.2
